replaced Integer-array with int-array

+++ src/main/java/mars/MarsLaunch.java
@@ -180,12 +180,12 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
          for (int i=0; i<dumpTriples.size(); i++) {
             String[] triple = (String[])dumpTriples.get(i);
             File file = new File(triple[2]);
-            Integer[] segInfo = MemoryDump.getSegmentBounds(triple[0]);
+            int[] segInfo = MemoryDump.getSegmentBounds(triple[0]);
          	// If not segment name, see if it is address range instead.  DPS 14-July-2008
             if (segInfo == null) {
                try {
                   String[] memoryRange = checkMemoryAddressRange(triple[0]);
-                  segInfo = new Integer[2];
+                  segInfo = new int[2];
                   segInfo[0] = Binary.stringToInt(memoryRange[0]); // low end of range
                   segInfo[1] = Binary.stringToInt(memoryRange[1]); // high end of range
                }    
@@ -208,12 +208,12 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
                continue;
             }
             try {
-               int highAddress = Globals.memory.getAddressOfFirstNull(segInfo[0].intValue(), segInfo[1].intValue())- Memory.WORD_LENGTH_BYTES;
-               if (highAddress < segInfo[0].intValue()) {
+               int highAddress = Globals.memory.getAddressOfFirstNull(segInfo[0], segInfo[1])- Memory.WORD_LENGTH_BYTES;
+               if (highAddress < segInfo[0]) {
                   out.println("This segment has not been written to, there is nothing to dump.");
                   continue;
                } 
-               format.dumpMemoryRange(file, segInfo[0].intValue(), highAddress); 
+               format.dumpMemoryRange(file, segInfo[0], highAddress);
             } 
                catch (FileNotFoundException e) {
                   out.println("Error while attempting to save dump, file " + file + " was not found!");
